requirement.txt
certifi==2020.4.5.2
chardet==3.0.4
idna==2.9
paho-mqtt==1.5.0
requests==2.23.0
urllib3==1.25.9
